Move pending button to left side of toolbar
