+++ dispatch/static/manager/src/js/components/ItemList/ItemListHeader.js
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { Button } from '@blueprintjs/core'
 
-import { Toolbar, ToolbarLeft, ToolbarRight, ToolbarMiddle } from '../Toolbar'
+import { Toolbar, ToolbarLeft, ToolbarRight } from '../Toolbar'
 import ItemListPagination from './ItemListPagination'
 import ItemListSearchBar from './ItemListSearchBar'
 
@@ -30,21 +30,15 @@ export default function ItemListHeader(props) {
         className='c-item-list__header__delete'
         onClick={() => props.actions.deleteItems(props.items.selected)}
         disabled={!props.items.selected.length}>Delete</Button>
+      {props.toolbarContent}
     </div>
   )
-
-  const toolbarMiddle = props.middleContent ? (
-    <ToolbarMiddle>
-      {props.middleContent}
-    </ToolbarMiddle>
-  ) : null
 
   return (
     <Toolbar alignLeft={true}>
       <ToolbarLeft>
         {props.actions.toggleAllItems ? toolbarLeft : null}
       </ToolbarLeft>
-      {toolbarMiddle}
       <ToolbarRight>
         {props.items.isLoaded && props.items.ids.length ? pagination : null}
         <ItemListSearchBar
