+++ dispatch/static/manager/src/js/components/ItemList/index.js
@@ -17,7 +17,7 @@ export default function ItemList(props) {
         currentPage={props.currentPage}
         totalPages={props.totalPages}
         createHandler={props.createHandler}
-        middleContent={props.toolbarMiddle}
+        toolbarContent={props.toolbarContent}
         actions={props.actions || {}} />
       <ItemListTable
         items={props.items}
