+++ dispatch/static/manager/src/js/components/Toolbar.js
@@ -28,9 +28,3 @@ export function ToolbarTitle(props) {
     <h3 className='c-toolbar__title'>{props.children}</h3>
   )
 }
-
-export function ToolbarMiddle(props) {
-  return (
-    <div className='c-toolbar__middle'>{props.children}</div>
-  )
-}
