+++ dispatch/static/manager/src/styles/components/toolbar.scss
@@ -22,6 +22,10 @@
   .pt-button {
     margin-right: 0.75rem;
   }
+
+  .pt-tag {
+    vertical-align: middle;
+  }
 }
 
 .c-toolbar__left {
